Rebase on Rocky Linux 9.3 image
The `cloud-init` shipped with Rocky Linux 9.4 has a bug that prevents
the correct pass of the `cloud-init` configuration at boot time via the
`smbios` qemu parameter. As a workaround, the process of image creation
follows these steps:
* Start from Rocky Linux 9.3
* Lock the version for the `cloud-init` package to prevent the update
  and the encounter of the bug
* Install NS8, which will also update to Rocky Linux 9.4
* Remove the lock on the `cloud-init` version

In this way, the images can be used on platforms that use the `smbios`
mechanism to pass the cloud-init configuration.